+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+# 2025-05-05
+- #2823 adds a task to monitor critial background tasks for errors/early exits and will trigger a rollup shutdown if any are detected.
+
 # 2025-05-01
 - #2821 simplifies several methods on `sov_metrics::MetricsTracker` by merging them into `MetricsTracker::submit` (and `submit_with_time`).